Add support for body text attachments
Add support for specifying text attachments to be dumped into an issue's
body. Do this to allow for output from previous DroneCI steps to be
added into the issue's message.

The version of the GitHub library used doesn't support attaching files
to the issue hence why I went with the approach of only supporting text
attachments and dumping the contents of the attachments in the issue
body.

Effort: 0:30 hours

Signed-off-by: Jason Rogena <jason@rogena.me>